+++ packages/material-ui/src/components/header.js
@@ -1,0 +1,19 @@
+import React, { PropTypes } from 'react';
+import getContext from 'recompose/getContext';
+
+const Header = getContext({
+  accounts: PropTypes.object,
+})(({ accounts }) =>
+  <div
+    style={{
+      fontSize: 32,
+      fontWeight: 400,
+      fontFamily: 'Roboto',
+      marginBottom: 10,
+    }}
+  >
+    {accounts.options().title}
+  </div>,
+  );
+
+export default Header;
