+++ packages/material-ui/src/resetPasswordComponents.js
@@ -5,6 +5,7 @@ import getContext from 'recompose/getContext';
 import Container from './Container';
 import Content from './Content';
 import FormError from './FormError';
+import Header from './components/header';
 import Avatar from './components/avatar';
 import PasswordField from './components/password';
 import SubmitButton from './components/submitButton';
@@ -17,21 +18,6 @@ const ResetPasswordFields = ({ children }) =>
 ResetPasswordFields.propTypes = {
   children: PropTypes.node,
 };
-
-const Header = getContext({
-  accounts: PropTypes.object,
-})(({ accounts }) =>
-  <div
-    style={{
-      fontSize: 32,
-      fontWeight: 400,
-      fontFamily: 'Roboto',
-      marginBottom: 10,
-    }}
-  >
-    {accounts.options().title}
-  </div>,
-  );
 
 const LoginButton = getContext({
   setFormType: PropTypes.func,
